:beetle: Fix config import
