+++ src/utils.js
@@ -17,23 +17,6 @@
 import Gio from 'gi://Gio';
 import GLib from 'gi://GLib';
 
-// Returns the given argument, except for "alpha", "beta", and "rc". In these cases -3,
-// -2, and -1 are returned respectively.
-function toNumericVersion(x) {
-  switch (x) {
-    case 'alpha':
-      return -3;
-    case 'beta':
-      return -2;
-    case 'rc':
-      return -1;
-  }
-  return x;
-}
-
-const Config               = imports.misc.config;
-const [GS_MAJOR, GS_MINOR] = Config.PACKAGE_VERSION.split('.').map(toNumericVersion);
-
 // This method can be used to write a message to GNOME Shell's log. This is enhances
 // the standard log() functionality by prepending the extension's name and the location
 // where the message was logged. As the extensions name is part of the location, you
@@ -46,33 +29,6 @@ export function debug(message) {
   stack.shift();
 
   log('[' + stack[0] + '] ' + message);
-}
-
-// Currently, the extension supports only one set of UI files. In the past, there were
-// three different sets for GTK3, GTK4, and Adwaita. This method returns the name of the
-// current UI directory. It's still here since it might be useful in the future.
-export function getUIDir() {
-  return 'adw';
-}
-
-// This method returns true if the current GNOME Shell version matches the given
-// arguments.
-export function shellVersionIs(major, minor) {
-  return GS_MAJOR == major && GS_MINOR == toNumericVersion(minor);
-}
-
-// This method returns true if the current GNOME Shell version is at least as high as the
-// given arguments. Supports "alpha" and "beta" for the minor version number.
-export function shellVersionIsAtLeast(major, minor) {
-  if (GS_MAJOR > major) {
-    return true;
-  }
-
-  if (GS_MAJOR == major) {
-    return GS_MINOR >= toNumericVersion(minor);
-  }
-
-  return false;
 }
 
 // Reads the contents of a file contained in the global resources archive. The data
